pull during cfetch
