+++ cargo-green/src/supergreen.rs
@@ -24,7 +24,6 @@ pub(crate) async fn main(green: Green, arg1: Option<&str>, args: Vec<String>) ->
     match arg1 {
         None | Some("-h" | "--help" | "-V" | "--version") => help(),
         Some("env") => envs(green, args),
-        Some("pull") => return pull(green).await,
         Some("push") => return push(green).await,
         Some(arg) => bail!("Unexpected supergreen command {arg:?}"),
     }
@@ -66,7 +65,7 @@ pub(crate) fn help() {
 
 Usage:
   cargo green supergreen env             Show used values
-  cargo green supergreen pull            Pulls images (respects $DOCKER_HOST)
+  cargo green fetch                      Pulls images (respects $DOCKER_HOST)
   cargo green supergreen push            Push cache image (all tags)
   cargo green supergreen -h | --help
   cargo green supergreen -V | --version
@@ -164,70 +163,4 @@ fn envs(green: Green, vars: Vec<String>) {
     if empty_vars {
         all.into_iter().for_each(|(var, o)| show(var, o.as_deref()));
     }
-}
-
-async fn pull(green: Green) -> Result<()> {
-    let imgs = [
-        (env::var(ENV_SYNTAX).ok(), green.syntax.as_str()),
-        (env::var(ENV_BASE_IMAGE).ok(), green.image.base_image.as_str()),
-        (env::var(ENV_BUILDER_IMAGE).ok(), green.builder_image.as_str()),
-        //FIXME: also pull xx image if needed
-    ]; // NOTE: we don't pull cache_image()
-
-    let mut to_pull = Vec::with_capacity(imgs.len());
-    for (user_input, img) in imgs {
-        let img = img.trim_start_matches("docker-image://");
-        let img = if img.contains('@') && user_input.map(|x| !x.contains('@')).unwrap_or_default() {
-            // Don't pull a locked image unless that's what's asked
-            // Otherwise, pull unlocked
-
-            // The only possible cases (user_input sets img)
-            // none + @ = trim
-            // none + _ = _
-            // s @  + @ = _
-            // s !  + @ = trim
-            trim_docker_image(img).expect("contains @")
-        } else {
-            img.to_owned()
-        };
-        to_pull.push(img);
-    }
-    pull_images(green, to_pull).await
-}
-
-#[must_use]
-fn trim_docker_image(x: &str) -> Option<String> {
-    let x = x.trim_start_matches("docker-image://");
-
-    let x =
-        if x.contains('@') { x.trim_end_matches(|c| c != '@').trim_end_matches('@') } else { x };
-
-    (!x.is_empty()).then(|| x.to_owned())
-}
-
-async fn pull_images(green: Green, to_pull: Vec<String>) -> Result<()> {
-    // TODO: nice TUI that handles concurrent progress
-    iter(to_pull.into_iter())
-        .map(|img| async {
-            println!("Pulling {img}...");
-            do_pull(&green, img).await
-        })
-        .buffer_unordered(10)
-        .try_collect()
-        .await
-}
-
-async fn do_pull(green: &Green, img: String) -> Result<()> {
-    let mut cmd = runner_cmd(green);
-    cmd.arg("pull").arg(&img);
-    let o = cmd
-        .spawn()
-        .map_err(|e| anyhow!("Failed to start {}: {e}", cmd.show()))?
-        .wait()
-        .await
-        .map_err(|e| anyhow!("Failed to call {}: {e}", cmd.show()))?;
-    if !o.success() {
-        bail!("Failed to pull {img}")
-    }
-    Ok(())
 }
